+++ doc/release-notes.md
@@ -30,6 +30,9 @@ None
 
 ## Modified functionality
 
+- The peer-to-peer network's block propagation logic has been improved to allow for lower-latency block propagation.
+  In particular, block downloads now request the latest block from up to 3 peers simultaneously so that the node has a
+  better chance of receiving the latest block as quickly as possible.
 - The `getnetworkinfo` RPC method results now include two new keys: `connections_in` and `connections_out`. These
   correspond to the current number of active inbound and outbound peer-to-peer connections, respectively.
 
@@ -43,7 +46,12 @@ None
 
 ## User interface changes
 
-None
+- The `getpeerinfo` RPC returns two new boolean fields, `bip152_hb_to` and
+  `bip152_hb_from`, that respectively indicate whether we selected a peer to be
+  in compact blocks high-bandwidth mode or whether a peer selected us as a
+  compact blocks high-bandwidth peer. High-bandwidth peers send new block
+  announcements via a `cmpctblock` message rather than the usual inv/headers
+  announcements. See BIP 152 for more details.
 
 ## Regressions